CONFIG: setup new package architecture with CircleCI, upgraded versions of deps, and modern config
This begins the work of a full re-write of the library. Developers will be required to be on the
latest versions of react, material ui, etc once the re-write is completed.

BREAKING CHANGE: This begins the work of a full re-write of the library. Developers will be required
to be on the latest versions of react, material ui, etc once the re-write is completed.